feat(core): cut the typed core and jeliyad over to protocol v2 (#165) (#166)
Implement the M1 typed-API-and-protocol-v2 milestone's daemon half.

#165 — typed core:
- projection.rs: pure conversions from internal/Iroh structures to the
  jeliya-api projections. Committed timeline events fold to typed
  jeliya-api Events (pos from the store Lamport clock, RFC 3339 <ts>,
  coupled kind-and-content), out-of-vocabulary status labels are omitted
  rather than reclassified, and unknown authors carry no fabricated role.
- typed.rs: TypedSupervisor serves every protocol-facing read as a typed
  jeliya-api value — rooms, members, timeline/archive paging, peers,
  fleet, files, pipes, invites, status history — plus the total typed
  dispatch table and the CoreError -> ApiError boundary mapping. No
  serde_json::Value escapes a protocol-facing projection.

#166 — v2-only engine and daemon:
- engine.rs: replace the 24-method v1 JSON dispatch with one typed
  execute(TypedCall) -> Result<TypedReply, ApiError> seam and a typed
  Push fan-out. PROTOCOL_VERSION/STORAGE_GENERATION = 2; no v1 facade.
- jeliyad serve.rs: run the codec's generation gate pre-upgrade (a v1
  client is refused 426 protocol_unsupported before any frame or
  dispatch), emit exactly one hello frame, and drive a v2-only loop —
  codec decode -> typed execute -> codec encode, with 4005/4007 closes
  and a backpressure gap push on lag.
- Quarantine jeliya-ffi from the active build (workspace exclude): it
  consumes the retired v1 string seam and is deleted under #202.
- Gate off the v1-coupled CI live suites (Dart/TS conformance, Node e2e,
  realnet harness) that drove the retired protocol; validate the v2
  conformance corpus shape instead.

Verified: workspace build, clippy -D warnings, and all tests pass; a
live v2 handshake (hello + typed subject.ensure/room.create/room.list/
room.timeline), typed Push::Event on message.send, daemon.stop
reply-first teardown, and a v1 client refused 426 pre-dispatch.

+++ crates/jeliyad/src/lifecycle.rs
@@ -38,7 +38,7 @@ pub struct Portfile {
     pub http: String,
     pub ws: String,
     pub version: String,
-    pub protocol: u32,
+    pub protocol: u64,
     pub data_dir: String,
     pub auth_token: String,
     pub started_at_ms: u64,
